requirements.txt
flask==2.3.3
openai==1.66.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Flask-WTF==1.1.1
pywebview==4.4.1
